Switch to a module for TypeId serialisation
